Server/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.2.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
contourpy==1.3.1
cryptography==42.0.5
cycler==0.12.1
daphne==4.1.2
Django==5.0.3
dnspython==2.6.1
flatbuffers==24.3.25
fonttools==4.57.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.0.8
mysql-connector-python==9.1.0
mysqlclient==2.2.4
namex==0.0.8
numpy==2.1.2
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
paho-mqtt==2.0.0
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
pyOpenSSL==24.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-etcd==0.4.5
pytz==2024.2
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
threadpoolctl==3.5.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.1.1
wheel==0.44.0
wrapt==1.16.0
zope.interface==6.2
